whisperFastAPIAsyncio-proyecto-tfg/requirements.txt
deep-translator==1.10.1
fastapi==0.95.2
Jinja2==3.1.2
openai-whisper==20230124
python-multipart==0.0.5
torch==2.0.1
uvicorn==0.22.0
yt_dlp==2023.3.4